Added in sun finder.
The sun will invariably overexpose any pixels it's present in, and some surrounding ones as well. Masking on overexposed pixels is a good start. It should also be round-ish. If there are too many overexposed pixels, it probably means clouds or the entire sky are overexposed, so that will be discarded.