Fog/cloud detection partly added in.
Fog and clouds have a common theme: very little high spatial frequency information. Using the Laplacian of the entire image is a easy and effective way to represent how blurry an image is. Note that this won't detect the presence of clouds in an otherwise normal image, it's for detecting when clouds are *all* that's visible