Added height estimation feature.
This uses the variation of Laplacian as a measure of detail again, and the idea that fine detail should fade as we get higher due to haze and turbulence. It's not very accurate though: limited sample data to train on (and I was only guessing at the actual height), wide range in ground structures and the ground/sky proportion in an image, and no consistency in the sensor/lens/conditions.